\name{as.data.frame.surveydata}
\alias{as.data.frame.surveydata}
\title{Coerces surveydata object to data.frame.}
\usage{
  \method{as.data.frame}{surveydata} (x, rm.pattern =
  FALSE)
}
\arguments{
  \item{x}{Surveydata object to coerce to class data.frame}

  \item{rm.pattern}{If TRUE removes \code{\link{pattern}}
  attributes from x}
}
\description{
  Coerces surveydata object to data.frame
}
\seealso{
  \code{\link{surveydata-package}}
}

