% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentext.R
\name{as_opentext_datatable}
\alias{as_opentext_datatable}
\title{Converts free format question text to datatable using the \code{DT} package.}
\usage{
as_opentext_datatable(data, q)
}
\arguments{
\item{data}{surveydata object}

\item{q}{Question}
}
\description{
Converts free format question text to datatable using the \code{DT} package.
}
\examples{
as_opentext_datatable(membersurvey, "Q33")
}
\seealso{
Other open text functions: 
\code{\link{print_opentext}()}
}
\concept{open text functions}
