% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{cbind.surveydata}
\alias{cbind.surveydata}
\title{Combines surveydata object by columns.}
\usage{
\method{cbind}{surveydata}(..., deparse.level = 1)
}
\arguments{
\item{...}{surveydata objects}

\item{deparse.level}{ignored}
}
\description{
Combines surveydata object by columns.
}
