% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{mutate.surveydata}
\alias{mutate.surveydata}
\alias{select.surveydata}
\alias{select_.surveydata}
\alias{filter.surveydata}
\alias{filter_.surveydata}
\alias{arrange.surveydata}
\alias{arrange_.surveydata}
\alias{summarize.surveydata}
\alias{summarize_.surveydata}
\title{Methods to support dplyr verbs.}
\usage{
mutate.surveydata(.data, ...)

select.surveydata(.data, ...)

select_.surveydata(.data, ...)

filter.surveydata(.data, ...)

filter_.surveydata(.data, ...)

arrange.surveydata(.data, ...)

arrange_.surveydata(.data, ...)

summarize.surveydata(.data, ...)

summarize_.surveydata(.data, ...)
}
\arguments{
\item{.data}{surveydata object or tbl passed to dplyr verb}

\item{...}{passed to dplyr verb}
}
\description{
Methods to support dplyr verbs.
}
