% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{fix_common_encoding_problems}
\alias{fix_common_encoding_problems}
\title{Fix common encoding problems when working with web imported data.}
\usage{
fix_common_encoding_problems(x, encoding = localeToCharset())
}
\arguments{
\item{x}{A character vector}

\item{encoding}{A character string describing the encoding of x.  Defaults to the current locale.  See also \code{\link[=iconvlist]{iconvlist()}}}
}
\description{
This function tries to resolve typical encoding problems when importing web data on Windows.
Typical problems occur with pound and emdash (-), especially when these originated in MS-Word.
}
\seealso{
Other Functions to clean data: 
\code{\link{encToInt}()},
\code{\link{fix_levels_01_spss}()},
\code{\link{has_dont_know}()},
\code{\link{intToEnc}()},
\code{\link{leveltest}},
\code{\link{remove_all_dont_know}()},
\code{\link{remove_dont_know}()}
}
\concept{Functions to clean data}
\keyword{encoding}
