% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{fix_levels_01_spss}
\alias{fix_levels_01_spss}
\alias{fix_levels_01_r}
\alias{fix_levels_01}
\title{Fix level formatting of all question with Yes/No type answers.}
\usage{
fix_levels_01_spss(dat)

fix_levels_01_r(dat)

fix_levels_01(dat, origin = c("R", "SPSS"))
}
\arguments{
\item{dat}{surveydata object}

\item{origin}{Either \code{R} or \code{SPSS}}
}
\description{
Fix level formatting of all question with Yes/No type answers.
}
\seealso{
Other Functions to clean data: 
\code{\link{encToInt}()},
\code{\link{fix_common_encoding_problems}()},
\code{\link{has_dont_know}()},
\code{\link{intToEnc}()},
\code{\link{leveltest}},
\code{\link{remove_all_dont_know}()},
\code{\link{remove_dont_know}()}
}
\concept{Functions to clean data}
\keyword{"clean}
\keyword{data"}
