\name{fixLevels01}
\alias{fixLevels01}
\title{Fix level formatting of all question with Yes/No type answers.}
\usage{
  fixLevels01(dat, origin = c("R", "SPSS"))
}
\arguments{
  \item{dat}{surveydata object}

  \item{origin}{Either "R" or "SPSS"}
}
\description{
  Fix level formatting of all question with Yes/No type
  answers.
}
\seealso{
  Other Functions to clean data: \code{\link{encToInt}},
  \code{\link{fixCommonEncodingProblems}},
  \code{\link{fixLevels01R}},
  \code{\link{fixLevels01SPSS}}, \code{\link{hasDK}},
  \code{\link{intToEnc}}, \code{\link{leveltestR}},
  \code{\link{leveltestSPSS}}, \code{\link{removeAllDK}},
  \code{\link{removeDK}}
}
\keyword{"clean}
\keyword{data"}

