% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydata.R
\name{is.surveydata}
\alias{is.surveydata}
\title{Tests whether an object is of class surveydata.}
\usage{
is.surveydata(x)
}
\arguments{
\item{x}{Object to check for being of class surveydata}
}
\description{
Tests whether an object is of class surveydata.
}
\seealso{
\link{surveydata-package}
}
