% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{lapply_names}
\alias{lapply_names}
\title{Applies function only to named elements of a list.}
\usage{
lapply_names(x, names, FUN, ...)
}
\arguments{
\item{x}{list}

\item{names}{character vector identifying which elements of the list to apply FUN}

\item{FUN}{function to apply.}

\item{...}{additional arguments passed to \code{FUN}}
}
\description{
This is useful to clean only some columns in a list (or \code{data.frame} or \code{surveydata} object). This is a simple wrapper around \code{\link[=lapply]{lapply()}} where only the named elements are changed.
}
\seealso{
Other Tools: 
\code{\link{question_order}()}
}
\concept{Tools}
