\name{leveltestR}
\alias{leveltestR}
\title{Fix level formatting of all question with Yes/No type answers.}
\usage{
  leveltestR(x)
}
\arguments{
  \item{x}{Surveyor data object}
}
\description{
  Fix level formatting of all question with Yes/No type
  answers.
}
\seealso{
  Other Functions to clean data: \code{\link{encToInt}},
  \code{\link{fixCommonEncodingProblems}},
  \code{\link{fixLevels01}}, \code{\link{fixLevels01R}},
  \code{\link{fixLevels01SPSS}}, \code{\link{hasDK}},
  \code{\link{intToEnc}}, \code{\link{leveltestSPSS}},
  \code{\link{removeAllDK}}, \code{\link{removeDK}}
}
\keyword{"clean}
\keyword{data"}

