% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_plot_title}
\alias{make_plot_title}
\title{Construct plot title from survey question.}
\usage{
make_plot_title(data, q)
}
\arguments{
\item{data}{surveydata object}

\item{q}{Question}
}
\description{
Construct plot title from survey question.
}
