% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydata-package.R
\docType{data}
\name{membersurvey}
\alias{membersurvey}
\title{Data frame with survey data of member satisfaction survey.}
\format{
data frame
}
\usage{
membersurvey
}
\description{
Data frame with survey data of member satisfaction survey.
}
\keyword{datasets}
