% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_varlabels}
\alias{merge_varlabels}
\title{Merges variable.labels attribute from two surveydata objects}
\usage{
merge_varlabels(dat1, dat2, new_names = union(names(dat1), names(dat2)))
}
\arguments{
\item{dat1}{surveydata object}

\item{dat2}{surveydata object}

\item{new_names}{A vector with names of the merged varlabels.  Defaults to the union of names of dat1 and dat2}
}
\description{
Merges variable labels from two data objects.  The labels from dat1 takes precedence.
}
\keyword{internal}
