\name{names_replace}
\alias{"names<-"}
\alias{names_replace}
\title{Updates names and variable.labels attribute of surveydata.}
\usage{
  names(x) <- value
}
\arguments{
  \item{x}{surveydata object}

  \item{value}{New names}
}
\description{
  Updates names and variable.labels attribute of
  surveydata.
}
\seealso{
  \code{\link{package-surveydata}},
  \code{\link{is.surveydata}}
}

