\name{pattern}
\alias{"pattern<-"}
\alias{pattern}
\alias{pattern<-}
\title{Returns and updates pattern attribute.}
\usage{
  pattern(x)

  pattern(x) <- value
}
\arguments{
  \item{x}{surveydata object}

  \item{value}{New value}
}
\description{
  \code{pattern} returns the pattern attribute of data, and
  \code{pattern<-} updates the attribute.
}
\seealso{
  \code{\link{surveydata-package}}
}

