% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-deprecated.R
\name{hasDK}
\alias{hasDK}
\alias{removeDK}
\alias{removeAllDK}
\alias{leveltestSPSS}
\alias{leveltestR}
\alias{fixLevels01SPSS}
\alias{fixLevels01R}
\alias{fixLevels01}
\alias{qOrder}
\alias{lapplyNames}
\title{Deprecated functions.}
\usage{
hasDK(...)

removeDK(...)

removeAllDK(...)

leveltestSPSS(...)

leveltestR(...)

fixLevels01SPSS(...)

fixLevels01R(...)

fixLevels01(...)

qOrder(...)

lapplyNames(...)
}
\description{
Deprecated functions.
}
