% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sats}
\alias{plot_sats}
\title{Plot satisfaction}
\usage{
plot_sats(data, q, fun = c("net", "top3", "top2"))
}
\arguments{
\item{data}{surveydata object}

\item{q}{Question}

\item{fun}{Aggregation function, one of \code{net} (compute net satisfaction score), \code{top3} (compute top 3 box score) and \code{top2} (compute top 2 box score)}
}
\description{
Plot satisfaction
}
\seealso{
Other plotting functions: \code{\link{plot_survey_question}},
  \code{\link{plot_yes_no}}
}
