% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_survey_question}
\alias{plot_survey_question}
\title{Plots single and  as multi-response questions.}
\usage{
plot_survey_question(data, q)
}
\arguments{
\item{data}{surveydata object}

\item{q}{Question}
}
\description{
Plots single and  as multi-response questions.
}
\seealso{
Other plotting functions: \code{\link{plot_sats}},
  \code{\link{plot_yes_no}}
}
