% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_yes_no}
\alias{plot_yes_no}
\title{Plot data in yes/no format.}
\usage{
plot_yes_no(data, q)
}
\arguments{
\item{data}{surveydata object}

\item{q}{Question}
}
\description{
Plot data in yes/no format.
}
\seealso{
Other plotting functions: \code{\link{plot_sats}},
  \code{\link{plot_survey_question}}
}
