% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{qOrder}
\alias{qOrder}
\title{Changes vector to ordered factor, adding NA levels if applicable.}
\usage{
qOrder(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Changes vector to ordered factor, adding NA levels if applicable.
}
\seealso{
Other Tools: \code{\link{lapplyNames}}
}
