\name{question_list}
\alias{question_list}
\title{Create list of questions in survey}
\usage{
  question_list(x, pattern = pattern(x))
}
\arguments{
  \item{x}{Object to coerce to surveydata}

  \item{pattern}{A regex pattern that defines the suffix of
  a question number}
}
\description{
  Create list of questions in survey
}
\keyword{tools}

