% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{question_order}
\alias{question_order}
\title{Changes vector to ordered factor, adding NA levels if applicable.}
\usage{
question_order(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Changes vector to ordered factor, adding NA levels if applicable.
}
\seealso{
Other Tools: \code{\link{lapply_names}}
}
\concept{Tools}
