% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{rm.attrs}
\alias{rm.attrs}
\title{Removes pattern and variable.labels from attributes list.}
\usage{
rm.attrs(x)
}
\arguments{
\item{x}{Surveydata object}
}
\description{
Removes pattern and variable.labels from attributes list.
}
\keyword{Internal}
