% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{rm.pattern}
\alias{rm.pattern}
\title{Removes pattern from attributes list.}
\usage{
rm.pattern(x)
}
\arguments{
\item{x}{Surveydata object}
}
\description{
Removes pattern from attributes list.
}
\keyword{Internal}
