% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{split_common_unique}
\alias{split_common_unique}
\title{Get common and unique text in question based on regex pattern identification.}
\usage{
split_common_unique(x, ptn = NULL)
}
\arguments{
\item{x}{A character vector}

\item{ptn}{A \code{\link[=regex]{regex()}} pattern that defines how the string should be split into common and unique elements}
}
\description{
Get common and unique text in question based on regex pattern identification.
}
\seealso{
Other Question functions: 
\code{\link{question_text_common}()},
\code{\link{question_text_unique}()},
\code{\link{question_text}()},
\code{\link{questions}()},
\code{\link{which.q}()}
}
\concept{Question functions}
\keyword{Questions}
