% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{splitCommonUnique}
\alias{splitCommonUnique}
\title{Get common and unique text in question based on regex pattern identification}
\usage{
splitCommonUnique(x, ptn = NULL)
}
\arguments{
\item{x}{A character vector}

\item{ptn}{A \code{\link[=regex]{regex()}} pattern that defines how the string should be split into common and unique elements}
}
\description{
Get common and unique text in question based on regex pattern identification
}
\seealso{
Other Question functions: \code{\link{qTextCommon}},
  \code{\link{qTextUnique}}, \code{\link{qText}},
  \code{\link{questions}}, \code{\link{which.q}}
}
\keyword{Questions}
