% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{survey_plot_title}
\alias{survey_plot_title}
\title{Construct plot title from the question text, wrapping at the desired width.}
\usage{
survey_plot_title(data, q, width = 50)
}
\arguments{
\item{data}{surveydata object}

\item{q}{Question}

\item{width}{Passed to \code{\link[=strwrap]{strwrap()}}}
}
\description{
This creates a plot title using \verb{[ggplot2::ggtitle()]}. The main title is string wrapped, and the subtitle is the number of observations in the data.
}
