% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{survey_plot_yes_no}
\alias{survey_plot_yes_no}
\title{Plot data in yes/no format.}
\usage{
survey_plot_yes_no(data, q)
}
\arguments{
\item{data}{surveydata object}

\item{q}{Question}
}
\description{
Plot data in yes/no format.
}
\seealso{
Other survey plotting functions: \code{\link{survey_plot_question}},
  \code{\link{survey_plot_satisfaction}}
}
