% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydata-deprecated.R
\name{hasDK}
\alias{hasDK}
\alias{removeDK}
\alias{removeAllDK}
\alias{leveltestSPSS}
\alias{leveltestR}
\alias{fixLevels01SPSS}
\alias{fixLevels01R}
\alias{fixLevels01}
\alias{qOrder}
\alias{lapplyNames}
\alias{fixCommonEncodingProblems}
\alias{qText}
\alias{qTextUnique}
\alias{qTextCommon}
\title{Deprecated functions.}
\usage{
hasDK(...)

removeDK(...)

removeAllDK(...)

leveltestSPSS(...)

leveltestR(...)

fixLevels01SPSS(...)

fixLevels01R(...)

fixLevels01(...)

qOrder(...)

lapplyNames(...)

fixCommonEncodingProblems(...)

qText(...)

qTextUnique(...)

qTextCommon(...)
}
\arguments{
\item{...}{passed to replacement function}
}
\description{
These functions have all been superseded with functions using \code{snake_case} function names.
\itemize{
\item \code{hasDK}: \code{\link[=has_dont_know]{has_dont_know()}}
\item \code{removeDK}: \code{\link[=remove_dont_know]{remove_dont_know()}}
\item \code{removeAllDK}: \code{\link[=remove_all_dont_know]{remove_all_dont_know()}}
\item \code{leveltestSPSS}: \code{\link[=leveltest_spss]{leveltest_spss()}}
\item \code{leveltestR}: \code{\link[=leveltest_r]{leveltest_r()}}
\item \code{fixLevels01SPSS}: \code{\link[=fix_levels_01_spss]{fix_levels_01_spss()}}
\item \code{fixLevels01R}: \code{\link[=fix_levels_01_r]{fix_levels_01_r()}}
\item \code{fixLevels01}: \code{\link[=fix_levels_01]{fix_levels_01()}}
\item \code{qOrder}: \code{\link[=question_order]{question_order()}}
\item \code{lapplyNames}: \code{\link[=lapply_names]{lapply_names()}}
\item \code{fixCommonEncodingProblems}: \code{\link[=fix_common_encoding_problems]{fix_common_encoding_problems()}}
}
}
\keyword{internal}
