\name{varlabels_replace}
\alias{"varlabels<-"}
\alias{varlabels_replace}
\title{Updates variable.labels attribute of surveyordata.}
\usage{
  varlabels(x) <- value
}
\arguments{
  \item{x}{surveydata object}

  \item{value}{New value}
}
\description{
  Updates variable.labels attribute of surveyordata.
}
\seealso{
  \code{\link{package-surveydata}}
}

