\name{as.data.frame.surveydata}
\alias{as.data.frame.surveydata}
\title{Coerces surveydata object to data.frame}
\usage{
  \method{as.data.frame}{surveydata} (x)
}
\arguments{
  \item{x}{Surveydata object to coerce to class data.frame}
}
\description{
  Coerces surveydata object to data.frame
}
\seealso{
  \code{\link{surveydata-package}}
}

