% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydata.R
\name{as.data.frame.surveydata}
\alias{as.data.frame.surveydata}
\alias{as.data.frame}
\title{Coerces surveydata object to data.frame.}
\usage{
\method{as.data.frame}{surveydata}(x, ..., rm.pattern = FALSE)
}
\arguments{
\item{x}{Surveydata object to coerce to class data.frame}

\item{...}{ignored}

\item{rm.pattern}{If TRUE removes \code{\link[=pattern]{pattern()}} attributes from x}
}
\description{
Coerces surveydata object to data.frame.
}
\seealso{
\link{surveydata-package}
}
