\name{as.surveydata}
\alias{as.surveydata}
\title{Coercion from and to surveydata.}
\usage{
  as.surveydata(x)
}
\arguments{
  \item{x}{Object to coerce to surveydata}
}
\description{
  Methods for creating "surveydata" objects, testing for
  class, and coercion from other objects.
}
\seealso{
  \code{\link{surveydata-package}},
  \code{\link{is.surveydata}}
}

