% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{dplyr-surveydata}
\alias{dplyr-surveydata}
\alias{mutate.surveydata}
\alias{as.tbl.surveydata}
\alias{select.surveydata}
\alias{filter.surveydata}
\alias{filter}
\alias{arrange.surveydata}
\alias{summarize.surveydata}
\alias{slice.surveydata}
\title{Methods to support dplyr verbs.}
\usage{
\method{mutate}{surveydata}(.data, ...)

as.tbl.surveydata(.data, ...)

\method{select}{surveydata}(.data, ...)

\method{filter}{surveydata}(.data, ...)

\method{arrange}{surveydata}(.data, ...)

\method{summarize}{surveydata}(.data, ...)

\method{slice}{surveydata}(.data, ...)
}
\arguments{
\item{.data}{surveydata object or tbl passed to dplyr verb}

\item{...}{passed to dplyr verb}
}
\description{
Methods to support dplyr verbs.
}
\details{
The surveydata exposes functionality to support some of the dplyr verbs, e.g. \code{\link[dplyr:filter]{dplyr::filter()}}
}
\examples{
library(dplyr)
membersurvey \%>\% as.tbl() \%>\% filter(Q2 == 2009)
}
\keyword{internal}
