% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{dplyr-surveydata}
\alias{dplyr-surveydata}
\alias{mutate_.surveydata}
\alias{select_.surveydata}
\alias{filter_.surveydata}
\alias{arrange_.surveydata}
\alias{summarize_.surveydata}
\alias{slice_.surveydata}
\title{Methods to support dplyr verbs.}
\usage{
mutate_.surveydata(.data, ...)

\method{select_}{surveydata}(.data, ...)

\method{filter_}{surveydata}(.data, ...)

\method{arrange_}{surveydata}(.data, ...)

\method{summarize_}{surveydata}(.data, ...)

\method{slice_}{surveydata}(.data, ...)
}
\arguments{
\item{.data}{surveydata object or tbl passed to dplyr verb}

\item{...}{passed to dplyr verb}
}
\description{
Methods to support dplyr verbs.
}
\examples{
library(dplyr)
membersurvey \%>\% as.tbl() \%>\% filter(Q2 == 2009)
}
\keyword{internal}
