% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{dplyr-surveydata}
\alias{dplyr-surveydata}
\alias{mutate.surveydata}
\alias{as.tbl.surveydata}
\alias{select.surveydata}
\alias{filter}
\alias{arrange.surveydata}
\alias{summarise.surveydata}
\alias{summarize.surveydata}
\alias{slice.surveydata}
\title{Methods to support dplyr verbs.}
\usage{
\method{mutate}{surveydata}(.data, ...)

\method{as.tbl}{surveydata}(.data, ...)

\method{select}{surveydata}(.data, ...)

\method{arrange}{surveydata}(.data, ...)

\method{summarise}{surveydata}(.data, ...)

\method{summarize}{surveydata}(.data, ...)

\method{slice}{surveydata}(.data, ...)
}
\arguments{
\item{.data}{\code{surveydata} object or \code{tbl} passed to \code{dplyr} verb}

\item{...}{passed to dplyr verb}
}
\description{
The \code{surveydata} package exposes functionality to support some of the \code{dplyr} verbs, e.g. \code{\link[dplyr:filter]{dplyr::filter()}}.  The computation is performed by \code{dplyr}, and the resulting object is of class \code{surveydata} (as well as the \code{dplyr} result).
}
\examples{
withr::with_package("dplyr", help = "dplyr", {
  membersurvey \%>\% 
    as.tbl() \%>\% 
    .[c("id", "Q1", "Q2")] \%>\% 
    filter(Q2 == 2009)
})


}
\keyword{internal}
