\name{extract}
\alias{"$<-.surveydata"}
\alias{"[.surveydata"}
\alias{[.surveydata}
\alias{extract}
\alias{replace}
\alias{subset}
\title{Extract or replace subsets of surveyordata.}
\usage{
  "x$name <- value"

  `x[i, j, ...]`
}
\arguments{
  \item{x}{surveydata object}

  \item{name}{Names of columns}

  \item{value}{New value}
}
\description{
  Extract or replace subsets of surveyordata, ensuring that
  the varlabels stay in synch.
}
\seealso{
  \code{\link{surveydata-package}}
}

