\name{extract}
\alias{"$<-"}
\alias{"$<-.surveydata"}
\alias{extract}
\alias{replace}
\title{Extract or replace subsets of surveydata.}
\usage{
  x$name <- value
}
\arguments{
  \item{x}{surveydata object}

  \item{name}{Names of columns}

  \item{value}{New value}
}
\description{
  Extract or replace subsets of surveydata, ensuring that
  the varlabels stay in synch.
}
\seealso{
  \code{\link{surveydata-package}}
}

