% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{Extract}
\alias{Extract}
\alias{[.surveydata}
\alias{[}
\alias{"[.surveydata"}
\alias{[<-.surveydata}
\alias{$<-.surveydata}
\alias{$<-}
\title{Extract or replace subsets of surveydata, ensuring that the varlabels stay synchronized.}
\usage{
\method{[}{surveydata}(x, i, j, drop = FALSE)

\method{[}{surveydata}(x, i, j) <- value

\method{$}{surveydata}(x, name) <- value
}
\arguments{
\item{x}{surveydata object}

\item{i}{row index}

\item{j}{column index}

\item{drop}{logical. Passed to \verb{[.data.frame}. Note that the default is \code{FALSE}.}

\item{value}{New value}

\item{name}{Names of columns}
}
\description{
The \code{surveydata} package makes it easy to extract specific questions from a surveydata object. Because survey data typically has question names like "Q1_a", "Q1_b", "Q1_c" the extract method for a \code{surveydata} object makes it easy to extract all columns by simply specifying "Q1" as the argument to the column index.
}
\details{
Extraction is similar to data frames, with three important exceptions:
\itemize{
\item The column argument \code{j} is evaluated using \code{\link[=which.q]{which.q()}} and will return all columns where the column names match the \code{\link[=pattern]{pattern()}}.
\item The \code{drop} argument is \code{FALSE}. Thus the result will always be a surveydata object, even if only a single column is returned.
\item All extraction methods retain the \code{pattern} and \code{varlabels} arguments.
}
}
\examples{

names(membersurvey)
head(membersurvey["Q1"])
head(membersurvey[c("Q1", "Q2")])
head(membersurvey[membersurvey$Q2=="2009", c("Q1", "Q2")])

# The pattern is used to extract columns

pattern(membersurvey)

grep("Q20", names(membersurvey), value=TRUE)
head(membersurvey["Q20"])
head(membersurvey["Q20_other"])

}
\seealso{
\link{surveydata-package}, \link{varlabels}
}
\keyword{internal}
