% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{has_dont_know}
\alias{has_dont_know}
\title{Tests whether levels contain "Don't know".}
\usage{
has_dont_know(x, dk = "Don't Know")
}
\arguments{
\item{x}{Character vector or factor}

\item{dk}{Character vector, containing search terms, e.g. \code{c("Don't know", "Don't Know")}}
}
\value{
TRUE or FALSE
}
\description{
Returns TRUE if x contains any instances of dk
}
\seealso{
Other Functions to clean data: 
\code{\link{encToInt}()},
\code{\link{fix_common_encoding_problems}()},
\code{\link{fix_levels_01_spss}()},
\code{\link{intToEnc}()},
\code{\link{leveltest}},
\code{\link{remove_all_dont_know}()},
\code{\link{remove_dont_know}()}
}
\concept{Functions to clean data}
\keyword{"clean}
\keyword{data"}
