% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{intToEnc}
\alias{intToEnc}
\title{Converts an integer vector to a character vector.}
\usage{
intToEnc(x, encoding = localeToCharset())
}
\arguments{
\item{x}{Integer vector}

\item{encoding}{A character string describing the encoding of x.  Defaults to the current locale.  See also \code{\link[=iconvlist]{iconvlist()}}}
}
\value{
A character vector
}
\description{
Conversion of integer vector to character vector.  The encoding of the character vector can be specified but defaults to the current locale.
}
\examples{
intToEnc(8212)
}
\seealso{
\code{\link[=iconv]{iconv()}}

Other Functions to clean data: 
\code{\link{encToInt}()},
\code{\link{fix_common_encoding_problems}()},
\code{\link{fix_levels_01_spss}()},
\code{\link{has_dont_know}()},
\code{\link{leveltest}},
\code{\link{remove_all_dont_know}()},
\code{\link{remove_dont_know}()}
}
\concept{Functions to clean data}
\keyword{encoding}
