\name{is.surveydata}
\alias{is.surveydata}
\title{Tests whether an object is of class surveydata.}
\usage{
  \method{is}{surveydata} (x)
}
\arguments{
  \item{x}{Object to check for being of class surveydata}
}
\description{
  Tests whether an object is of class surveydata.
}
\seealso{
  \code{\link{package-surveydata}}
}

