% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{leveltest}
\alias{leveltest}
\alias{leveltest_spss}
\alias{leveltest_r}
\title{Fix level formatting of all question with Yes/No type answers.}
\usage{
leveltest_spss(x)

leveltest_r(x)
}
\arguments{
\item{x}{surveydata object}
}
\description{
Fix level formatting of all question with Yes/No type answers.
}
\seealso{
Other Functions to clean data: 
\code{\link{encToInt}()},
\code{\link{fix_common_encoding_problems}()},
\code{\link{fix_levels_01_spss}()},
\code{\link{has_dont_know}()},
\code{\link{intToEnc}()},
\code{\link{remove_all_dont_know}()},
\code{\link{remove_dont_know}()}
}
\concept{Functions to clean data}
\keyword{"clean}
\keyword{data"}
