% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{leveltestR}
\alias{leveltestR}
\title{Fix level formatting of all question with Yes/No type answers.}
\usage{
leveltestR(x)
}
\arguments{
\item{x}{surveydata object}
}
\description{
Fix level formatting of all question with Yes/No type answers.
}
\seealso{
Other Functions to clean data: \code{\link{encToInt}},
  \code{\link{fixCommonEncodingProblems}},
  \code{\link{fixLevels01R}},
  \code{\link{fixLevels01SPSS}}, \code{\link{fixLevels01}},
  \code{\link{hasDK}}, \code{\link{intToEnc}},
  \code{\link{leveltestSPSS}}, \code{\link{removeAllDK}},
  \code{\link{removeDK}}
}
\keyword{"clean}
\keyword{data"}
