\name{names}
\alias{"names<-"}
\alias{names}
\title{Updates names and variable.labels attribute of surveydata.}
\usage{
  names(x) <- value
}
\arguments{
  \item{value}{New names}

  \item{x}{surveydata object}
}
\description{
  Updates names and variable.labels attribute of
  surveydata.
}
\seealso{
  \code{\link{surveydata-package}},
  \code{\link{is.surveydata}}
}

