% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveydata.R
\name{names<-.surveydata}
\alias{names<-.surveydata}
\alias{names<-}
\title{Updates names and variable.labels attribute of surveydata.}
\usage{
\method{names}{surveydata}(x) <- value
}
\arguments{
\item{x}{surveydata object}

\item{value}{New names}
}
\description{
Updates names and variable.labels attribute of surveydata.
}
\seealso{
\code{\link[=surveydata-package]{surveydata-package()}}, \code{\link[=is.surveydata]{is.surveydata()}}
}
\keyword{internal}
