% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pattern}
\alias{pattern}
\alias{pattern<-}
\title{Returns and updates pattern attribute.}
\usage{
pattern(x)

pattern(x) <- value
}
\arguments{
\item{x}{surveydata object}

\item{value}{New value}
}
\description{
The pattern attribute contains information about the separator character used to name sub-questions in the data.  Survey software typically makes use of underscores to distinguish sub-questions in a grid of questions, e.g. "Q4_1", "Q4_2", "Q4_3", "Q4_other". The function \code{\link[=pattern]{pattern()}} returns the \code{pattern} attribute, and \link{pattern<-} updates the attribute.
}
\examples{
# Extract the pattern from membersurvey

oldptn <- pattern(membersurvey)
oldptn

# The pattern is used to extract columns

names(membersurvey)
grep("Q20", names(membersurvey), value=TRUE)

head(membersurvey["Q20"])
head(membersurvey["Q20_other"])

# Define a new pattern

pattern(membersurvey) <- list(sep="_", exclude="")
head(membersurvey["Q20"])

# Reset original pattern

pattern(membersurvey) <- oldptn
rm(oldptn)
}
\seealso{
\code{\link[=as.surveydata]{as.surveydata()}}, \code{\link[=which.q]{which.q()}}

Other Attribute functions: 
\code{\link{varlabels}()}
}
\concept{Attribute functions}
\keyword{internal}
