\name{qText}
\alias{qText}
\title{Returns question text.}
\usage{
  qText(x, Q)
}
\arguments{
  \item{x}{A surveydata object}

  \item{Q}{The question id, e.g. Q4}
}
\description{
  Given a question id, e.g. "Q4", returns question text
}
\seealso{
  Other Questions: \code{\link{qTextCommon}},
  \code{\link{qTextUnique}}, \code{\link{questions}},
  \code{\link{which.q}}
}

