\name{splitCommonUnique}
\alias{splitCommonUnique}
\title{Get common and unique text in question based on regex pattern identification}
\usage{
  splitCommonUnique(x, ptn = NULL)
}
\arguments{
  \item{x}{A character vector}

  \item{ptn}{A \code{\link{regex}} pattern that defines how
  the string should be split into common and unique
  elements}
}
\description{
  Get common and unique text in question based on regex
  pattern identification
}

