% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{survey_plot_question}
\alias{survey_plot_question}
\title{Plots single and  as multi-response questions.}
\usage{
survey_plot_question(data, q)
}
\arguments{
\item{data}{surveydata object}

\item{q}{Question}
}
\description{
Plots single and  as multi-response questions.
}
\examples{
question_text(membersurvey)

survey_plot_question(membersurvey, "Q2")
survey_plot_yes_no(membersurvey, "Q2")
survey_plot_satisfaction(membersurvey, "Q14")

}
\seealso{
Other survey plotting functions: 
\code{\link{survey_plot_satisfaction}()},
\code{\link{survey_plot_yes_no}()}
}
\concept{survey plotting functions}
