\name{varlabels}
\alias{"varlabels<-"}
\alias{varlabels}
\alias{varlabels<-}
\title{Returns and updates variable.labels attribute of data.}
\usage{
  varlabels(x)
}
\arguments{
  \item{x}{surveydata object}
}
\description{
  \code{varlabels} returns the variable.labels attribute of
  data, and \code{varlabels(x) <- value} updates the
  attribute.
}
\seealso{
  \code{\link{surveydata-package}}
}

