\name{varlabels}
\alias{"varlabels<-"}
\alias{varlabels}
\alias{varlabels<-}
\title{Returns and updates variable.labels attribute of data.}
\usage{
  varlabels(x)

  varlabels(x) <- value
}
\arguments{
  \item{x}{surveydata object}

  \item{value}{New value}
}
\description{
  \code{varlabels} returns the variable.labels attribute of
  data, and \code{varlabels<-} updates the attribute.
}
\seealso{
  \code{\link{surveydata-package}}
}

